# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"Leave Management API","description":"Employees, leave policies, holidays, leave requests, approvals and notifications. All /api endpoints need HTTP Basic auth (employee email + password).","version":"v1"},"servers":[{"url":"http://localhost:8080","description":"Local"}],"security":[{"basicAuth":[]}],"tags":[{"name":"1. Employees","description":"Accounts, roles and reporting lines"},{"name":"2. Leave policies","description":"Yearly quota and carry-forward per grade and leave type"},{"name":"3. Holidays","description":"Company holidays (not counted as leave days)"},{"name":"4. Leave requests","description":"Apply, view, cancel, balance and yearly summary"},{"name":"5. Approvals","description":"Manager decisions (tasks are created automatically)"},{"name":"6. Notifications","description":"Notifications of the logged-in user"}],"paths":{"/api/policies/{id}":{"put":{"tags":["2. Leave policies"],"summary":"Update policy quota (HR)","operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateLeavePolicyRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LeavePolicy"}}}}}},"delete":{"tags":["2. Leave policies"],"summary":"Delete policy (HR)","operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"},"example":9}],"responses":{"200":{"description":"OK"}}}},"/api/notifications/{id}/read":{"put":{"tags":["6. Notifications"],"summary":"Mark one notification read","operationId":"markRead","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Notification"}}}}}}},"/api/notifications/read-all":{"put":{"tags":["6. Notifications"],"summary":"Mark all my notifications read","operationId":"markAllRead","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"integer","format":"int32"}}}}}}}},"/api/leave-requests/{id}/cancel":{"put":{"tags":["4. Leave requests"],"summary":"Cancel leave (pending, or approved and not started)","operationId":"cancel","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"},"example":3}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LeaveRequest"}}}}}}},"/api/employees/{id}":{"get":{"tags":["1. Employees"],"summary":"Get employee (self, manager or HR)","operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"},"example":3}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeResponse"}}}}}},"put":{"tags":["1. Employees"],"summary":"Update employee (HR)","operationId":"update_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"},"example":3}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateEmployeeRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeResponse"}}}}}},"delete":{"tags":["1. Employees"],"summary":"Deactivate employee (HR)","operationId":"deactivate","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"},"example":4}],"responses":{"200":{"description":"OK"}}}},"/api/employees/me/password":{"put":{"tags":["1. Employees"],"summary":"Change my password","operationId":"changePassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/approvals/{id}/decide":{"put":{"tags":["5. Approvals"],"summary":"Approve or reject (assigned approver only)","operationId":"decide","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"},"example":1},{"name":"decision","in":"query","required":true,"schema":{"type":"string","enum":["APPROVED","REJECTED"]},"example":"APPROVED"},{"name":"comments","in":"query","required":false,"schema":{"type":"string"},"example":"Enjoy your time"}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApprovalTask"}}}}}}},"/api/policies":{"get":{"tags":["2. Leave policies"],"summary":"List policies","operationId":"getAll","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LeavePolicy"}}}}}}},"post":{"tags":["2. Leave policies"],"summary":"Create policy (HR)","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LeavePolicyRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LeavePolicy"}}}}}}},"/api/policies/evaluate":{"post":{"tags":["2. Leave policies"],"summary":"Evaluate a request against the policy (nothing is saved)","operationId":"evaluate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LeaveEvaluationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LeaveEvaluationResponse"}}}}}}},"/api/leave-requests":{"post":{"tags":["4. Leave requests"],"summary":"Apply for leave (as the logged-in employee)","operationId":"apply","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplyLeaveRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LeaveRequest"}}}}}}},"/api/holidays":{"get":{"tags":["3. Holidays"],"summary":"List holidays of a year","operationId":"getByYear","parameters":[{"name":"year","in":"query","required":false,"schema":{"type":"integer","format":"int32"},"example":2026}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Holiday"}}}}}}},"post":{"tags":["3. Holidays"],"summary":"Add holiday (HR)","operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HolidayRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Holiday"}}}}}}},"/api/employees":{"get":{"tags":["1. Employees"],"summary":"List employees (HR: all, MANAGER: own team)","operationId":"list","parameters":[{"name":"page","in":"query","description":"Zero-based page index (0..N)","required":false,"schema":{"type":"integer","default":0,"minimum":0}},{"name":"size","in":"query","description":"The size of the page to be returned","required":false,"schema":{"type":"integer","default":20,"minimum":1}},{"name":"sort","in":"query","description":"Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.","required":false,"schema":{"type":"array","default":["id,ASC"],"items":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PagedModelEmployeeResponse"}}}}}},"post":{"tags":["1. Employees"],"summary":"Create employee (HR)","operationId":"create_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateEmployeeRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeResponse"}}}}}}},"/api/notifications":{"get":{"tags":["6. Notifications"],"summary":"My notifications (newest first)","operationId":"getMine","parameters":[{"name":"unreadOnly","in":"query","required":false,"schema":{"type":"boolean","default":false},"example":true},{"name":"page","in":"query","description":"Zero-based page index (0..N)","required":false,"schema":{"type":"integer","default":0,"minimum":0}},{"name":"size","in":"query","description":"The size of the page to be returned","required":false,"schema":{"type":"integer","default":20,"minimum":1}},{"name":"sort","in":"query","description":"Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.","required":false,"schema":{"type":"array","default":["id,DESC"],"items":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PagedModelNotification"}}}}}}},"/api/leave-requests/{id}":{"get":{"tags":["4. Leave requests"],"summary":"Get leave request (owner, manager or HR)","operationId":"getById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LeaveRequest"}}}}}}},"/api/leave-requests/summary/{employeeId}":{"get":{"tags":["4. Leave requests"],"summary":"Yearly summary: rejected and pending counts","operationId":"getSummary","parameters":[{"name":"employeeId","in":"path","required":true,"schema":{"type":"integer","format":"int64"},"example":3},{"name":"year","in":"query","required":true,"schema":{"type":"integer","format":"int32"},"example":2026}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LeaveSummaryResponse"}}}}}}},"/api/leave-requests/my":{"get":{"tags":["4. Leave requests"],"summary":"My leave requests","operationId":"getMine_1","parameters":[{"name":"page","in":"query","description":"Zero-based page index (0..N)","required":false,"schema":{"type":"integer","default":0,"minimum":0}},{"name":"size","in":"query","description":"The size of the page to be returned","required":false,"schema":{"type":"integer","default":20,"minimum":1}},{"name":"sort","in":"query","description":"Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.","required":false,"schema":{"type":"array","default":["startDate,DESC"],"items":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PagedModelLeaveRequest"}}}}}}},"/api/leave-requests/employee/{employeeId}":{"get":{"tags":["4. Leave requests"],"summary":"Leave requests of an employee (self, manager or HR)","operationId":"getByEmployee","parameters":[{"name":"employeeId","in":"path","required":true,"schema":{"type":"integer","format":"int64"},"example":3},{"name":"page","in":"query","description":"Zero-based page index (0..N)","required":false,"schema":{"type":"integer","default":0,"minimum":0}},{"name":"size","in":"query","description":"The size of the page to be returned","required":false,"schema":{"type":"integer","default":20,"minimum":1}},{"name":"sort","in":"query","description":"Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.","required":false,"schema":{"type":"array","default":["startDate,DESC"],"items":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PagedModelLeaveRequest"}}}}}}},"/api/leave-requests/balance":{"get":{"tags":["4. Leave requests"],"summary":"My leave balance","operationId":"getMyBalance","parameters":[{"name":"year","in":"query","required":false,"schema":{"type":"integer","format":"int32"},"example":2026}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LeaveBalanceResponse"}}}}}}}},"/api/leave-requests/balance/{employeeId}":{"get":{"tags":["4. Leave requests"],"summary":"Leave balance of an employee (self, manager or HR)","operationId":"getBalance","parameters":[{"name":"employeeId","in":"path","required":true,"schema":{"type":"integer","format":"int64"},"example":3},{"name":"year","in":"query","required":false,"schema":{"type":"integer","format":"int32"},"example":2026}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LeaveBalanceResponse"}}}}}}}},"/api/employees/me":{"get":{"tags":["1. Employees"],"summary":"My profile","operationId":"me","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeResponse"}}}}}}},"/api/approvals/{id}":{"get":{"tags":["5. Approvals"],"summary":"Get approval task (approver or HR)","operationId":"getById_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApprovalTask"}}}}}}},"/api/approvals/pending":{"get":{"tags":["5. Approvals"],"summary":"My pending approvals","operationId":"getMyPending","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ApprovalTask"}}}}}}}},"/api/holidays/{id}":{"delete":{"tags":["3. Holidays"],"summary":"Delete holiday (HR)","operationId":"delete_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"},"example":2}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"UpdateLeavePolicyRequest":{"type":"object","properties":{"annualQuota":{"type":"integer","format":"int32","example":10,"maximum":365,"minimum":0},"carryForwardLimit":{"type":"integer","format":"int32","example":3,"maximum":365,"minimum":0}},"required":["annualQuota","carryForwardLimit"]},"LeavePolicy":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"leaveType":{"type":"string","enum":["CASUAL","SICK","EARNED"]},"grade":{"type":"string"},"annualQuota":{"type":"integer","format":"int32"},"carryForwardLimit":{"type":"integer","format":"int32"}}},"Notification":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"recipientId":{"type":"integer","format":"int64"},"message":{"type":"string"},"read":{"type":"boolean"},"createdAt":{"type":"string","format":"date-time"}}},"LeaveRequest":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"employeeId":{"type":"integer","format":"int64"},"leaveType":{"type":"string","enum":["CASUAL","SICK","EARNED"]},"startDate":{"type":"string","format":"date"},"endDate":{"type":"string","format":"date"},"days":{"type":"integer","format":"int32"},"status":{"type":"string","enum":["PENDING","APPROVED","REJECTED","REJECTED_BY_POLICY","CANCELLED"]},"reason":{"type":"string"},"policyRemarks":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"version":{"type":"integer","format":"int64"}}},"UpdateEmployeeRequest":{"type":"object","properties":{"name":{"type":"string","example":"Ravi Kumar Singh","maxLength":100,"minLength":0},"grade":{"type":"string","example":"L1","minLength":1,"pattern":"L1|L2|L3"},"role":{"type":"string","example":"EMPLOYEE","minLength":1,"pattern":"EMPLOYEE|MANAGER|HR"},"managerId":{"type":"integer","format":"int64","example":2}},"required":["grade","name","role"]},"EmployeeResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"email":{"type":"string"},"grade":{"type":"string"},"role":{"type":"string","enum":["EMPLOYEE","MANAGER","HR"]},"managerId":{"type":"integer","format":"int64"},"joiningDate":{"type":"string","format":"date"},"active":{"type":"boolean"}}},"ChangePasswordRequest":{"type":"object","properties":{"currentPassword":{"type":"string","example":"Ravi@1234","minLength":1},"newPassword":{"type":"string","example":"Ravi@5678","maxLength":72,"minLength":8}},"required":["currentPassword","newPassword"]},"ApprovalTask":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"leaveRequestId":{"type":"integer","format":"int64"},"approverId":{"type":"integer","format":"int64"},"status":{"type":"string","enum":["PENDING","APPROVED","REJECTED","CANCELLED"]},"comments":{"type":"string"},"decidedAt":{"type":"string","format":"date-time"},"version":{"type":"integer","format":"int64"}}},"LeavePolicyRequest":{"type":"object","properties":{"leaveType":{"type":"string","example":"EARNED","minLength":1,"pattern":"CASUAL|SICK|EARNED"},"grade":{"type":"string","example":"L3","minLength":1,"pattern":"L1|L2|L3"},"annualQuota":{"type":"integer","format":"int32","example":21,"maximum":365,"minimum":0},"carryForwardLimit":{"type":"integer","format":"int32","example":8,"maximum":365,"minimum":0}},"required":["annualQuota","carryForwardLimit","grade","leaveType"]},"LeaveEvaluationRequest":{"type":"object","properties":{"grade":{"type":"string","example":"L1","minLength":1},"leaveType":{"type":"string","example":"CASUAL","minLength":1,"pattern":"CASUAL|SICK|EARNED"},"requestedDays":{"type":"integer","format":"int32","example":12,"minimum":1},"alreadyUsedDays":{"type":"integer","format":"int32","example":0,"minimum":0},"carriedForwardDays":{"type":"integer","format":"int32","example":3,"minimum":0}},"required":["alreadyUsedDays","grade","leaveType","requestedDays"]},"LeaveEvaluationResponse":{"type":"object","properties":{"approved":{"type":"boolean"},"reason":{"type":"string"}}},"ApplyLeaveRequest":{"type":"object","properties":{"leaveType":{"type":"string","example":"CASUAL","minLength":1,"pattern":"CASUAL|SICK|EARNED"},"startDate":{"type":"string","format":"date","example":"2026-09-28"},"endDate":{"type":"string","format":"date","example":"2026-10-02"},"reason":{"type":"string","example":"Family function","maxLength":255,"minLength":0}},"required":["endDate","leaveType","startDate"]},"HolidayRequest":{"type":"object","properties":{"date":{"type":"string","format":"date","example":"2026-10-02"},"name":{"type":"string","example":"Gandhi Jayanti","maxLength":100,"minLength":0}},"required":["date","name"]},"Holiday":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"holidayDate":{"type":"string","format":"date"},"name":{"type":"string"}}},"CreateEmployeeRequest":{"type":"object","properties":{"name":{"type":"string","example":"Ravi Kumar","maxLength":100,"minLength":0},"email":{"type":"string","format":"email","example":"ravi@company.com","maxLength":150,"minLength":0},"password":{"type":"string","example":"Ravi@1234","maxLength":72,"minLength":8},"grade":{"type":"string","example":"L1","minLength":1,"pattern":"L1|L2|L3"},"role":{"type":"string","example":"EMPLOYEE","minLength":1,"pattern":"EMPLOYEE|MANAGER|HR"},"managerId":{"type":"integer","format":"int64","example":2},"joiningDate":{"type":"string","format":"date","example":"2023-04-01"}},"required":["email","grade","name","password","role"]},"PageMetadata":{"type":"object","properties":{"size":{"type":"integer","format":"int64"},"number":{"type":"integer","format":"int64"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int64"}}},"PagedModelNotification":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/Notification"}},"page":{"$ref":"#/components/schemas/PageMetadata"}}},"LeaveSummaryResponse":{"type":"object","properties":{"rejectedCount":{"type":"integer","format":"int64"},"unapprovedCount":{"type":"integer","format":"int64"}}},"PagedModelLeaveRequest":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/LeaveRequest"}},"page":{"$ref":"#/components/schemas/PageMetadata"}}},"LeaveBalanceResponse":{"type":"object","properties":{"leaveType":{"type":"string","enum":["CASUAL","SICK","EARNED"]},"annualQuota":{"type":"integer","format":"int32"},"carriedForward":{"type":"integer","format":"int32"},"used":{"type":"integer","format":"int32"},"pending":{"type":"integer","format":"int32"},"remaining":{"type":"integer","format":"int32"}}},"PagedModelEmployeeResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/EmployeeResponse"}},"page":{"$ref":"#/components/schemas/PageMetadata"}}}},"securitySchemes":{"basicAuth":{"type":"http","description":"Username = employee email, password = employee password","scheme":"basic"}}}}